Add language describing the removal of confirm_mac
We're replacing the confirm_mac with a trailing MAC as in other ZRTP
messages.

This has the advantage of binding the MAC for the encrypted section to
the Message Type Block.  It also removes a one-off MAC definition from
the protocol.

This does mean the MAC is no longer placed at a non-fixed offset in th
message, but this is already true of the DHPart1 and DHPart2 messages.