Add a fingerprint block to Confirm
We'd like for there to be a way to use the ZRTP SA to authenticate an
outside pair of keys.  If the keys were guaranteed to be known ahead
of time, we could conceivably use auxsecret for this as in:

  https://tools.ietf.org/html/draft-johnston-rtcweb-zrtp-02

However we'd like to support cases where one or both keys may not be
known by the other party at the time of the ZRTP exchange.  Therefore
we'll put the public key fingerprints in the Confirm message along
with a type block to identify the intended use for each fingerprint.