Extend the ZID to 128 bits
UUIDs are now commonly 128 bits long.  Phil set the ZID to 96 bits by
observing that MAC addresses are 48 bits long and that people seem
satisfied that space will never be exhausted.  As MAC addresses live
in a managed space, he figured that 96 bits should be long enough for
an unmanaged space.

He was probably right there, but the rest of the world decided to be
32 bits more conservative, so we'd like to pick the more commonly
accepted length of 128-bits for a UUID, which is what the ZID is.