Add surnames for Keccak/SHA-3 authors
xml2rfc v2.9.6 (at least) has a bug when generating HTML output when
encountering the last author of a reference with multiple authors
where no prior author had a surname or an organizational affiliation.

In xml2rfc, in `write_reference_list` in `html.py`, the lack of a
surname or organization causes a loop to continue before setting a
`last` variable, which is then dereferenced unconditionally while
processing the last author.

Earlier versions of xml2rfc (circa 2015) did not have this bug.

The Keccak/SHA-3 references had multiple authors but lacked surnames,
thereby triggering the bug.  This commit adds surnames for the SHA-3
authors to avoid it.